Add files via upload
source : Kaggle

This Dataset contains the number of vehicle sales realized in a certain period. The number of sales was kept for every month. We will forecast sales numbers for the next 6 months using MLP (Multi-layer Perceptron). We want to forecast for the next 6 months. However, we will not be able to see how successful the results we produced are, as we do not have actual sales data to compare our predicted result with. Therefore, to measure the success of the forecast, i set aside the last 6 months of the data as test data. i leave the rest as a train. we will predict the next months from the train data and compare it with the real values in the test data.